requirements.txt
flask
python-dotenv
flask_sqlalchemy
openai